added a welcome screen (this will be fragmented later)
