+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,15 @@
         </activity>
 
         <activity
+            android:name=".DayViewActivity"
+            android:parentActivityName=".MainActivity">
+            <!-- Parent activity meta-data to support 4.0 and lower -->
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".MainActivity" />
+        </activity>
+
+        <activity
             android:name=".AddSubjectActivity"
             android:parentActivityName=".MainActivity"
             android:theme="@style/EditorTheme">
